Updated transitions, added visual settings
Added toggles for VSync and Fullscreen, and a resolution dropdown. Blocked user input on transitions (while fading in/out). I used my usual node set-up of an overall World node, so the tree is a bit different from the original but much of the code should be possible to translate over!